added builder method to User, changed name of LoginView to AuthView, updated AuthController to reflect new name
