--- src/View/LoginView.java
@@ -1,72 +1,0 @@
-package View;
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-
-public class LoginView {
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(() -> createLoginView());
-    }
-
-    private static void createLoginView() {
-        JFrame loginFrame = new JFrame("Login");
-        loginFrame.setSize(300, 150);
-        loginFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-
-        JPanel loginPanel = new JPanel();
-        loginFrame.add(loginPanel);
-        placeComponents(loginPanel);
-
-        loginFrame.setLocationRelativeTo(null);
-        loginFrame.setVisible(true);
-    }
-
-    private static void placeComponents(JPanel panel) {
-        panel.setLayout(null);
-
-        JLabel userLabel = new JLabel("Username:");
-        userLabel.setBounds(10, 10, 80, 25);
-        panel.add(userLabel);
-
-        JTextField userText = new JTextField(20);
-        userText.setBounds(100, 10, 160, 25);
-        panel.add(userText);
-
-        JLabel passwordLabel = new JLabel("Password:");
-        passwordLabel.setBounds(10, 40, 80, 25);
-        panel.add(passwordLabel);
-
-        JPasswordField passwordText = new JPasswordField(20);
-        passwordText.setBounds(100, 40, 160, 25);
-        panel.add(passwordText);
-
-        JButton loginButton = new JButton("Login");
-        loginButton.setBounds(10, 80, 80, 25);
-        panel.add(loginButton);
-
-        JButton cancelButton = new JButton("Cancel");
-        cancelButton.setBounds(100, 80, 80, 25);
-        panel.add(cancelButton);
-
-        loginButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                String username = userText.getText();
-                String password = new String(passwordText.getPassword());
-
-                // Implement your authentication logic here
-                System.out.println("Username: " + username);
-                System.out.println("Password: " + password);
-            }
-        });
-
-        cancelButton.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                System.exit(0);
-            }
-        });
-    }
-}
